rename quiz.js to quiz_session.js
