+++ src/quiz/quiz_command_handler.js
@@ -4,7 +4,7 @@ const QuizJSON      = require("../../" + questionsFile);
 const Bot           = require("../hopson_bot");
 const Util          = require("../misc/util");
 const CommandHandlerBase = require("../command_handler_base")
-const QuizSession = require("./quiz")
+const QuizSession = require("./quiz_session")
 const Config        = require('../../data/config.json');
 
 const fs            = require('fs');
